requirements.txt
python-dotenv
prompt_toolkit
openai
